Refactored metric test datasets
